fix: move test, service, repository about Member
